+++ posthog-node/CHANGELOG.md
@@ -1,7 +1,7 @@
 # NEXT
 
-- `flushAsync` is removed with `flush` now being an async method.
-- Fixed an issue where `shutdownAsync` would potentially exit early if a flush was already in progress
+- `flushAsync` and `shutdownAsnc` are removed with `flush` and `shutdown` now being the async methods.
+- Fixed an issue where `shutdown` would potentially exit early if a flush was already in progress
 - Flushes will now try to flush up to `maxBatchSize` (default 100) in one go
 
 # 4.0.0-beta.1 - 2024-03-04